Feat : set dtos
add dtos as records for each model. that later will have 3 dtos : requestDTO(have the necessary attribute to create object from the request data from the client) && embeddedDTO(have the pure attribute of the model so we can inject it in other object) && responseDTO(the object that we will return to the client).